+++ src/runtime/src/value_iterator.rs
@@ -215,14 +215,6 @@ impl Iterator for ValueIterator {
     }
 }
 
-pub fn is_iterable(value: &Value) -> bool {
-    use Value::*;
-    matches!(
-        value,
-        Range(_) | List(_) | Tuple(_) | Map(_) | Str(_) | Iterator(_)
-    )
-}
-
 pub fn make_iterator(value: &Value) -> Result<ValueIterator, ()> {
     use Value::*;
     let result = match value {